# openapi.yaml
openapi: 3.1.0
info:
  title: SlipScribe API
  version: 1.0.0
  description: |
    SlipScribe backend API for receipt capture, OCR extraction, line-item parsing, and spending insights.
    
    ## Authentication
    All endpoints require Bearer token authentication via `Authorization: Bearer <token>` header.
    
  contact:
    name: SlipScribe Team
  license:
    name: MIT

servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.slipscribe.com/api/v1
    description: Production

security:
  - bearerAuth: []

paths:
  /receipts/upload:
    post:
      summary: Upload a receipt image or PDF
      operationId: uploadReceipt
      tags:
        - receipts
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Receipt image (JPEG, PNG, HEIC) or PDF
                merchant_name:
                  type: string
                  description: Optional pre-filled merchant name
                purchase_date:
                  type: string
                  format: date
                  description: Optional pre-filled purchase date
      responses:
        '201':
          description: Receipt uploaded and queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'

  /receipts:
    get:
      summary: List user receipts
      operationId: listReceipts
      tags:
        - receipts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed, manual_review]
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: merchant
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of receipts
          content:
            application/json:
              schema:
                type: object
                properties:
                  receipts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Receipt'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

  /receipts/{receipt_id}:
    get:
      summary: Get receipt details
      operationId: getReceipt
      tags:
        - receipts
      parameters:
        - $ref: '#/components/parameters/ReceiptId'
      responses:
        '200':
          description: Receipt details with line items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    patch:
      summary: Update receipt fields
      operationId: updateReceipt
      tags:
        - receipts
      parameters:
        - $ref: '#/components/parameters/ReceiptId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                merchant_name:
                  type: string
                purchase_date:
                  type: string
                  format: date
                subtotal:
                  type: number
                  format: decimal
                tax:
                  type: number
                  format: decimal
                total:
                  type: number
                  format: decimal
                currency:
                  type: string
                  default: USD
      responses:
        '200':
          description: Receipt updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /receipts/{receipt_id}/line-items:
    get:
      summary: Get receipt line items
      operationId: getReceiptLineItems
      tags:
        - line-items
      parameters:
        - $ref: '#/components/parameters/ReceiptId'
      responses:
        '200':
          description: List of line items
          content:
            application/json:
              schema:
                type: object
                properties:
                  line_items:
                    type: array
                    items:
                      $ref: '#/components/schemas/LineItem'
        '404':
          $ref: '#/components/responses/NotFound'

  /receipts/{receipt_id}/line-items/{line_item_id}:
    patch:
      summary: Update a line item
      operationId: updateLineItem
      tags:
        - line-items
      parameters:
        - $ref: '#/components/parameters/ReceiptId'
        - name: line_item_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                quantity:
                  type: number
                  format: decimal
                unit_price:
                  type: number
                  format: decimal
                line_total:
                  type: number
                  format: decimal
                discount_flag:
                  type: boolean
      responses:
        '200':
          description: Line item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineItem'
        '404':
          $ref: '#/components/responses/NotFound'

  /receipts/{receipt_id}/reprocess:
    post:
      summary: Reprocess a receipt (re-run OCR and extraction)
      operationId: reprocessReceipt
      tags:
        - receipts
      parameters:
        - $ref: '#/components/parameters/ReceiptId'
      responses:
        '202':
          description: Reprocessing job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [queued]
        '404':
          $ref: '#/components/responses/NotFound'

  /search:
    get:
      summary: Search receipts and line items
      operationId: searchReceipts
      tags:
        - search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query (merchant name, item name, or full text)
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchResult'
                  total:
                    type: integer

  /dashboard/monthly:
    get:
      summary: Get monthly spending dashboard
      operationId: getMonthlyDashboard
      tags:
        - dashboard
      parameters:
        - name: month
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
            example: '2026-03'
          description: Month in YYYY-MM format (defaults to current month)
      responses:
        '200':
          description: Monthly dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyDashboard'

  /insights/overspend:
    get:
      summary: Get overspend alerts
      operationId: getOverspendInsights
      tags:
        - insights
      responses:
        '200':
          description: Overspend insights
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/OverspendAlert'

  /exports/csv:
    get:
      summary: Export receipts to CSV
      operationId: exportCSV
      tags:
        - exports
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ReceiptId:
      name: receipt_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    Receipt:
      type: object
      required:
        - id
        - user_id
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        merchant_name:
          type: string
          nullable: true
        purchase_date:
          type: string
          format: date
          nullable: true
        subtotal:
          type: number
          format: decimal
          nullable: true
        tax:
          type: number
          format: decimal
          nullable: true
        total:
          type: number
          format: decimal
          nullable: true
        currency:
          type: string
          default: USD
        status:
          type: string
          enum: [pending, processing, completed, failed, manual_review]
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ReceiptDetail:
      allOf:
        - $ref: '#/components/schemas/Receipt'
        - type: object
          properties:
            images:
              type: array
              items:
                $ref: '#/components/schemas/ReceiptImage'
            line_items:
              type: array
              items:
                $ref: '#/components/schemas/LineItem'
            validation_warnings:
              type: array
              items:
                type: string

    ReceiptImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        page_index:
          type: integer
        width:
          type: integer
        height:
          type: integer

    LineItem:
      type: object
      required:
        - id
        - receipt_id
        - name
      properties:
        id:
          type: string
          format: uuid
        receipt_id:
          type: string
          format: uuid
        name:
          type: string
        quantity:
          type: number
          format: decimal
          default: 1
        unit_price:
          type: number
          format: decimal
          nullable: true
        line_total:
          type: number
          format: decimal
        discount_flag:
          type: boolean
          default: false
        raw_text:
          type: string
          nullable: true
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    SearchResult:
      type: object
      properties:
        type:
          type: string
          enum: [receipt, line_item]
        receipt_id:
          type: string
          format: uuid
        line_item_id:
          type: string
          format: uuid
          nullable: true
        merchant_name:
          type: string
        purchase_date:
          type: string
          format: date
        total:
          type: number
          format: decimal
          nullable: true
        matched_text:
          type: string
        score:
          type: number
          format: float

    MonthlyDashboard:
      type: object
      properties:
        month:
          type: string
          example: '2026-03'
        total_spent:
          type: number
          format: decimal
        receipt_count:
          type: integer
        by_category:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              amount:
                type: number
                format: decimal
              receipt_count:
                type: integer
        by_merchant:
          type: array
          items:
            type: object
            properties:
              merchant_name:
                type: string
              amount:
                type: number
                format: decimal
              receipt_count:
                type: integer
        daily_trend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              amount:
                type: number
                format: decimal

    OverspendAlert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category:
          type: string
          nullable: true
        current_month_total:
          type: number
          format: decimal
        baseline_average:
          type: number
          format: decimal
        percent_change:
          type: number
          format: float
        severity:
          type: string
          enum: [info, warning, critical]
        message:
          type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: object

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Not found

    PayloadTooLarge:
      description: File too large
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: File size exceeds 10MB limit